Added Historical Data Requests. I made it so all Datetimes are stored in the stack to avoid manual memory allocation.   Same with CorrelationId.   Same with Service.   Same with ObjectType. Added some utility functions to Datetime that are not DLL visible.   This allowed me to clean up some of the Datetime math I have. Made numValues and numElements size_t rather than int. Element get Value/Element As String now returns const char * rather than char *. Fixed a bug with my strncmp calls.   I wasn't accounting for the null-terminator in the length argument.   Some of my strncmp calls suffered from copy-and-paste errors.  I fixed those too.
Former-commit-id: 2cae28de8d8c510eec68900d88ccc97a4a0178f6

+++ BEmu_cpp/headers/BloombergTypes/Message.h
@@ -43,7 +43,7 @@ namespace BEmu
 			DLL_EXPORT CorrelationId correlationId() const;
 			DLL_EXPORT Name messageType() const;
 			DLL_EXPORT virtual const char* topicName() const;
-			DLL_EXPORT virtual int numElements() const;
+			DLL_EXPORT virtual size_t numElements() const;
 			//DLL_EXPORT Service service() const;
 			DLL_EXPORT virtual const Element asElement() const;
