+++ BEmu_cpp/headers/IntradayBarRequest/ElementBarTickData.h
@@ -37,8 +37,8 @@ namespace BEmu
 				~ElementBarTickData();
 
 				virtual Name name() const;
-				virtual int numValues() const;
-				virtual int numElements() const;
+				virtual size_t numValues() const;
+				virtual size_t numElements() const;
 		
 				virtual bool isArray() const;
 				virtual bool isComplexType() const;
