+++ BEmu_cpp/headers/IntradayBarRequest/ElementBarTickDataArray.h
@@ -31,8 +31,8 @@ namespace BEmu
 				~ElementBarTickDataArray();
 
 				virtual Name name() const;
-				virtual int numValues() const;
-				virtual int numElements() const;
+				virtual size_t numValues() const;
+				virtual size_t numElements() const;
 		
 				virtual bool isArray() const;
 				virtual bool isComplexType() const;
