+++ BEmu_cpp/headers/IntradayBarRequest/RequestIntradayBarElementStringArray.h
@@ -34,8 +34,8 @@ namespace BEmu
 				std::vector<std::string*> values();
 
 				virtual Name name() const;
-				virtual int numValues() const;
-				virtual int numElements() const;
+				virtual size_t numValues() const;
+				virtual size_t numElements() const;
 		
 				virtual bool isArray() const;
 				virtual bool isComplexType() const;
